Tests/LibWeb/Text/expected/css/parse-nested-pseudo-selectors-as-selectors.txt

This text should look green  Color is: "rgb(0, 128, 0)", should be "rgb(0, 128, 0)"
